Add two tests.
The first shows that upon calling a function, superglobals are initialized before the parameters are processed. This fails for us.

The second shows that array accesses, field accesses and symbol table accesses, while they don't deal with their indexes in identical ways, they are otherwise the same concept. Therefore it may be desirable to replace Rvalue with ARRAY_INDEX for array_indices in Array_access and Assign_array.