Update the version of libtool. This uses the latest git version of libtool, instead of the old 1.5.x version. The new version is 2.2.7a-ish.
The reason for the new version is that it fixes the problem of running the combination of libtool, gdb and plugins. It's a little unusual to not use a release version, but there is cause in this case. I submitted the bug report for this, and had to test it anyway. Since I had to update to 2.2.6 anywy to test it, it seemed a waste not to commit it, especially since it took some time.

Annoyingly, this requires that running gdb be done with:
	./libtool --mode=execute gdb ...
instead of
	./libtool gdb ...

However, it seems we can shorten it to:
	./libtool e gdb ...
which isnt bad at all.